requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
Flask-SocketIO==5.5.1
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-socketio==5.13.0
aiohttp==3.12.13
aiofiles==24.1.0
requests==2.32.4
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
email-validator==2.2.0
Werkzeug==3.1.3

